ref: update jwt ver and guard roles
