+++ src/auth/guard/role.guard.ts
@@ -7,8 +7,7 @@ import { RequestWithUser } from '../interfaces/auth.interfaces';
 export const RoleGuard = (roles: Role[]): Type<CanActivate> => {
   class RoleGuardMixin implements CanActivate {
     canActivate(context: ExecutionContext) {
-      const request = context.switchToHttp().getRequest<RequestWithUser>();
-      const user = request.user;
+      const { user } = context.switchToHttp().getRequest<RequestWithUser>();
       if (user?.roles.some((role) => roles.includes(role))) {
         return true;
       } else {
