+++ src/auth/strategy/jwt.strategy.ts
@@ -1,22 +1,26 @@
-import { UsersService } from '@app/users/users.service';
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { JwtTokenConfType } from '../interfaces/auth.enum';
 import { TokenPayload } from '../interfaces/auth.interfaces';
+import { AuthUserService } from '../services/auth-user.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor(private readonly configService: ConfigService, private readonly usersService: UsersService) {
+  constructor(private readonly configService: ConfigService, private readonly authUserService: AuthUserService) {
     super({
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: true,
+      ignoreExpiration: false,
       secretOrKey: configService.get(JwtTokenConfType.access).tokenSecretKey,
     });
   }
 
   async validate(payload: TokenPayload) {
-    return this.usersService.getById(payload.userId);
+    try {
+      return await this.authUserService.validateUser(payload.userId);
+    } catch (e) {
+      throw e;
+    }
   }
 }
